requirements.txt
flask==3.0.3
line-bot-sdk==3.14.1
gspread==6.1.2
google-auth==2.29.0
google-api-python-client==2.126.0
openpyxl==3.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
